Using an inteface for SwrveEventsManager so can be swapped easily in testing.
